Changed screen handlers behaviour. Added bgimages printing function.
Now modes receive the handlers and only exist one.
Now screen handlers constructor do only video, memory and background initialization.
Screen handlers have methods to init the mode of the screen and to switch
from a mode to another.
Added draw background image function to screens handlers to draw the images
acording to the screen and game mode.
Adapted all the modes to the new behaviour.