Added palette and tick and cross images to bg images.
Added black and gray colors in an extra row to the images in order to
render the fonts, which color is searched in the palette.
Added tick and cross images below in the background images to have its
colors in the palette. Now simply print a region out of the screen of
the bitmap wherever.
Added and moved functions to graphics namespace.